Update Polish Translation (#229)
* Update Polish Translation - UpdateJson

UpdateJson in this context doesn't mean 'update the json file', but 'a link to a JSON file that contains info so manager can update the module'.
It is also a field described in the Magisk documentation, so it's best to leave the original form.

* Update Polish Translation - Export logs

Mostly a cosmetic change, should fit better with the other options in the settings menu